+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v0.7.5...HEAD)
+## [Unreleased](https://github.com/fivetran/go-fivetran/compare/v0.7.6...HEAD)
 
 ## Added
 Supported the following Fivetran API endpoints:
@@ -16,8 +16,8 @@ Supported the following Fivetran API endpoints:
 - [Delete webhook](https://fivetran.com/docs/rest-api/webhooks#deletewebhook)
 - [Retrieve the list of webhooks](https://fivetran.com/docs/rest-api/webhooks#retrievethelistofwebhooks)
 - [Test webhook](https://fivetran.com/docs/rest-api/webhooks#testwebhook)
-
-## [0.7.5](https://github.com/fivetran/go-fivetran/compare/v0.7.4...v0.7.5) Not Published
+ 
+## [0.7.6](https://github.com/fivetran/go-fivetran/compare/v0.7.5...v0.7.6)
 
 ## Fixed
 - DBT Transformations: `paused` field supported.